WIP: Change MultiBot to store bots keyed by their username, allow modules to be enabled/disabled dynamically, and allow bots to be started and stopped dynamically.
This is not tied in to /newbot at all, so that module is now completely broken. However, tests are all passing because metabot.modules.conftest uses a mock MultiBot that does not follow the real MultiBot. Fixing that (which will make fixing /newbot much easier) will be invasive enough I want to separate these changes.