Find and install all modules (and loadable packages) present in the metabot/modules/ directory.
The new events module is proving really tricky for a number of reasons, not the least of which being the mess caused by the partial deprecation of the oauth2client package (in favor of google-auth). I'm leaning toward using https://packaging.python.org/guides/packaging-namespace-packages/#pkgutil-style-namespace-packages to split it into a separate metabot-events package, installing it as lib/pythonX/site-packages/metabot/modules/events/. This would require either some kind of manual configuration of modules (maybe in a new section of config/multibot.json) or having the bot find its modules dynamically. For now, the latter seems simplest.