Simple date humanization logic.
This uses the timezone of the system where the bot is running, meaning users should invoke the bot as something like:

TZ=US/Pacific python -m metabot

to force all events to be returned in Pacific time. (This also means pytest should be invoked the same way!) A followup should probably make this a per-group and/or per-user setting (and at the same time fix the timezone for testing).