Add two more settings to the moderator module: maxeventsdays and maxeventscount. /events in a group now searches for maxeventsdays days, instead of hard coding to 6, and truncates the results at maxeventscount, now defaulting to 10.
Testing adminui via test_events is becoming too bulky, so a followup should introduce a separate test_adminui.

The existing fields (calendars, greeting, and timezone) have been pretty self-explanatory, but these new fields are not very intuitive at all. A followup should improve the documentation of all fields (both the in-editor text and the labels in the listing).

Fixes #7.