Initialize modules right before initializing individual bots.
If a module has a `modinit` function, call it (passing in the MultiBot instance).

Don't perform any periodic events (such as polling the MultiCalendar) via MultiBot; every module will be responsible for adding its own periodics via `modinit`, on whatever schedule they want. (This means that a module that wants to refresh data from Google Drive once an hour might be scheduled to run before or after, nondeterministically, a module that wants to use that data, so this might end up being changed again in the future.)

The events module will now handle polling the MultiCalendar. I think the creation of the MultiCalendar should also be handled by the events module, but it's a little tricky given how adminui needs to see the MultiCalendar-generated calcodes, so I'll leave MultiBot.calendars (and MultiBot.multical) in place for now.

See #3 and #25.