In a group chat, allow /events to include events through the end of the `maxeventsdays`th day, rather than exactly 60 * 60 * 24 * `maxeventsdays` seconds.
This means using a value of 0 now means "events occurring from now until the end of today" instead of just "events occurring right now". If it ever becomes desirable to allow the bot to only report events that are occuring right now, we can remove the + 1 in the `period` calculation and change the default of `maxeventsdays` from 6 to 7.

Fixes #28.