When a user follows a more-details link from /events (in a group or otherwise), have Prev/Next only navigate the calendar that event is on, and always use the timezone used for the presentation the more-details link was generated from.
For example, if a user is in @myfungroup, which is set to display events from Calendar A in timezone US/Eastern, and types /events, the user can click on the event's location to see more details about the event in US/Eastern, and Prev will take them to the previous event on Calendar A. This happens regardless of whether the user has other calendars enabled via /events set, even if Calendar A isn't set or the user hasn't set any calendars at all; and similarly even if they have set their timezone to US/Pacific or haven't set one at all.

The similarity between the now, effectively, two different versions of /events (in private) might be too confusing. One quick followup might be to do something like:

         if ':' in eventid and timezone:
             suffix = ' ' + timezone
             calcodes = eventid.split(':', 1)[0]
    +        msg.path('/events', 'Events')
    +        msg.path(calcodes, calcodes)
         else:

(or preferably actually look up the calendar's friendly name), so there will be both a visual indicator they are in a special calendar view and MessageBuilder will automatically add a Back button to take them to the normal /events. For now, the UI replaces the Settings/Current button with a My Events button that switches you into manual /events mode.

Fixes #34.