Change the timezone adminui widget to prompt for country code rather than extracting it from language_code.
US and Canadian users no longer view the US/ and Canada/ prefixes, but the full entries for their countries (from America/ and Pacific/). This is suboptimal UI-wise, but is presumably more correct. The list of zones is now provided in zone.tab, rather than alphabetical, order, so it should make more sense when navigating. Additionally, only the city name is provided, which should be more human friendly; and the zone.tab comment (if any) is also tacked on.

Fixes #52 (for now).