Allow group admins to access moderator settings via /mod (fixes #10).
For now, just literally expose moderator.admin to users via a new /mod command.

Also for now, don't document /mod in /help. A followup will probably add it only when it's relevant (as admin.modhelp does).

There's a slight risk here:  We poll for group admins once an hour, so theoretically an abusive admin could be demoted quickly but still have up to an hour to cause more problems by messing with the bot's group settings.