Store the `records` dict in CONFDIR/daily.pickle.
I really wanted to store this in multibot.conf (implicitly in CONFDIR/daily.yaml, or possibly even in CONFIDR/bots.yaml somewhere), but a number of issues—including the fact that I'm storing floats, tuples, and lists with dicts as children—made that really awkward. This might one day be revisited, but for now this will allow me to move ahead without the bot immediately losing track of its daily announcements every time it's upgraded (or otherwise restarted).

Fixes #64.