Eliminate Menu.select(..., create=True) in favor of allowing Menu.add(field=None) to catch anything not explicitly specified.
Maintain Menu.fields in sorted order, if only to guarantee field=None is the last entry checked. This could have been implemented as explicitly forcing field=None to the end while preserving the rest of the natural order, but I don't think it's really worth allowing non-alphabetical fields in the UI.

See #66.