Display Google Calendar "flair" banners in front of /events and daily announcements if the first event listed matches on of its patterns.
Note that the selection of flair (both whether flair is displayed or not, and which image is used) is currently decided permanently during the initial daily announcement. If the first event is removed, pushed down by a new event, or changes its summary/description to no longer match the chosen image, the image will remain in place; only the text will be updated.

See #74.