Display NWS alerts in groups with a daily announcement configured and at least one event (with a geocode-able location) in the next 90 days.
See https://github.com/nmlorg/metabot/issues/80#issuecomment-560515034.

Test running this for the past few days, it turns out there's a lot of noise due to both (1) api.weather.gov's intermittent unavailability and (2) NWS frequently updating alerts to tweak the wording, estimates, or affected areas. The latter sometimes causes alerts to be re-posted even if the 250-character blurb doesn't change, simply because the way they update an alert is to cancel it and issue a brand new one in its place (so the alert id, which is included in the URL, changes).

A followup should probably include some kind of diffing behavior, possibly literally calling _quick_diff, but tweaking it to at least show 250 characters of context, and possibly allowing it to actually find multiple changes.