Somewhat hacky initial implementation of #83.
Bot admins can now customize the image used for a given event when it is the first event in either /events or a daily announcement. Admins can choose to customize either the specific event (by local_id) or the event's series (by title match).

Right now, admins can not cleanly back out of the new Customize UI because `/events admin` without an event id is not a valid path; this will be fixed in a followup.

Unrelated, but while I'm thinking about it, it may make sense for a followup to change things to use https://core.telegram.org/bots/api#sendmediagroup instead of the current https://core.telegram.org/bots/api#sendphoto so images from all (or the next X? the next Y with photos?) upcoming events are included.