Add a basic event RSVP/notes feature (see #110).
Note that this does not implement the personal reminders mentioned in #110 (see #48). This only allows people to mark that they are/maybe/are not going to an event, and/or to add a freeform note to the event.

Everywhere (including daily announcements), a count of yesses and maybes is added to the event title, and events with any number of notes has a "note" emoji in the "more details" text:

  My Test Event  👍 2  🤔 1
  📝  🔜 ²ʰ¹⁰ᵐ Sat 17ᵗʰ, 12–1ᵖᵐ

The full list of yesses (and notes) is included when a user gets full details (in a private chat):

  My Test Event  👍 2  🤔 1
  📝  🔜 ²ʰ²ᵐ Sat 17ᵗʰ, 12–1ᵖᵐ

  This is a test event.

  This is the end of the description in Google Calendar.

  Attending:
  • Some User — Test note.
  • Other User

  (I have you down as a maybe 🤔.)

  Notes:
  • user123456789 — Note from unknown user.
  [Definitely going 👍] [Definitely not 🙁] [Add note 📝]