Add event banners to /events in private messages (see #114).
This change causes /events in private messages (as well as when clicking on "more details" links) to send two messages:  A new message that only contains the event's banner is sent immediately before the existing reply.

The existing reply is then modified to include the message_id of the image message as well as the URL (or reference) used in a hidden tg://meta/ link.

Whenever the user clicks Prev or Next, if the new event has a different banner, the image-only message is first edited to display the new banner, then the text message is edited as usual. (This causes both an increase in the already-perceivable lag before Prev/Next buttons take effect, as well as a visual jump, where the image changes, followed by a perceivable lag, then the text changing.)

For now, I'm using Google's "planmyday" event illustration as the image for events that don't have images themselves. A followup may switch to using something like https://placehold.co/600x200.png?text=Event+Title, and/or maybe actually synthesizing a calendar (though "anti-AI sentiment" is very high right now, so navigating that is obviously tricky).

Release as 0.3.30.